+++ src/components/inputs/plansAndPromotions/coupons.tsx
@@ -1,104 +1,326 @@
-import { SyntheticEvent, useCallback, useState } from "react";
+import { SyntheticEvent, useCallback, useEffect, useState } from "react";
 import TextInput from "../textInput";
 import LoadingButton from "../../button";
 import Select from "../select";
 import { useAppDispatch } from "../../../stores/hooks";
-import { addPriceTypeToList } from "../../../stores/apiData/price-type-lists";
-import { addCouponsToList } from "../../../stores/apiData/coupons-lists";
+import {
+  addCouponsToList,
+  replaceCouponsInList,
+} from "../../../stores/apiData/coupons-lists";
 import Search from "../search";
+import DateInput from "../dateInput";
+import useAxios from "../../../useHooks/useAxios";
+import useGetCoupon from "../../../services-hooks/useGetCoupon";
+import {
+  openSnackbar,
+  snackBar,
+} from "../../../stores/appFunctionality/snackbar";
 
-export default function AddNewCoupon({ setOpen }: { setOpen: Function }) {
+export default function AddNewCoupon({
+  setOpen,
+  id,
+}: {
+  setOpen: Function;
+  id?: string;
+}) {
   const dispatch = useAppDispatch();
+  const axios = useAxios();
+
+  const { data, isLoading, isFailed, setIsFailed, retryFunction } =
+    useGetCoupon({ id });
+
   const [code, setCode] = useState("");
+  const [startDate, setStartDate] = useState("");
+  const [endDate, setEndDate] = useState("");
+  const [apartmentApplicability, setApartmentApplicability] = useState("");
+  const [assignedApartments, setAssignedApartments] = useState([]);
+  const [userApplicability, setUserApplicability] = useState("");
+  const [assignCustomers, setAssignedCustomers] = useState("");
   const [type, setType] = useState("");
-  const [totalOrPercent, setTotalOrPercent] = useState("");
-  const [value, setValue] = useState("");
-  const [customer, setCustomer] = useState("");
+  const [validity, setValidity] = useState("");
+  const [isReusable, setIsReusable] = useState("");
+  const [price, setPrice] = useState("");
+  const [percentage, setPercentage] = useState("");
+  const [minAmount, setMinAmount] = useState("");
+  const [maxAmount, setMaxAmount] = useState("");
 
   const [isSubmitting, setIsSubmitting] = useState(false);
 
-  const submitPricing = useCallback(
+  // populate data authomatically
+  useEffect(() => {
+    if (id && data) {
+      const {
+        code,
+        start_date,
+        end_date,
+        applicable_to_shortlet,
+        applicable_to_user,
+        type,
+        validity,
+        is_reusable,
+        price,
+        percentage,
+        minimum_amount,
+        maximum_amount,
+      } = data;
+      setCode(code);
+      setStartDate(start_date);
+      setEndDate(end_date);
+      setApartmentApplicability(applicable_to_shortlet);
+      setUserApplicability(applicable_to_user);
+      setType(type);
+      setValidity(validity);
+      setIsReusable(is_reusable === 1 ? "yes" : "no");
+      setPrice(String(price));
+      setPercentage(String(percentage));
+      setMinAmount(String(minimum_amount));
+      setMaxAmount(String(maximum_amount));
+    }
+  }, [id, data]);
+
+  const generateCouponCode = useCallback(async () => {
+    const generatedCode = Math.random()
+      .toString(36)
+      .substring(2, 5 + 2)
+      .toLocaleUpperCase();
+    setCode(generatedCode);
+  }, []);
+
+  const createCoupon = useCallback(
     async (e: SyntheticEvent) => {
       e.preventDefault();
-      //   const payload = {
-      //     code,
-      //     type,
-      //     totalOrPercent,
-      //     value,
-      //     customer
-      //   };
+      const payload = {
+        code: code,
+        start_date: startDate,
+        end_date: endDate,
+        applicable_to_shortlet: apartmentApplicability, // all or specific
+        applicable_to_user: userApplicability, // all or specific
+        type: type, //percentage or price
+        validity: validity, //temporary or permanent
+        is_reusable: isReusable, // yes or no
+        applicable_shortlets: assignedApartments, // required if applicable_to_shortlet is specific
+        applicable_users: assignCustomers, //required if applicable_to_user is specific
+        price: price, //required if type is price
+        percentage: percentage, //required if type is percentage
+        minimum_amount: minAmount,
+        maximum_amount: maxAmount,
+      };
+
       try {
+        let response;
+        if (id) {
+          response = await axios.put(`admin/coupon/${id}`, payload);
+          const dataset = response?.data?.data;
+          dispatch(replaceCouponsInList(dataset));
+        } else {
+          response = await axios.post("/admin/coupon", payload);
+          const dataset = response?.data?.data;
+          dispatch(addCouponsToList(dataset));
+        }
+        console.log({ response });
+        setOpen(false);
         dispatch(
-          addCouponsToList({
-            id: "random",
-            name: "new-coupon",
-            type: type,
-            validityDates: "today",
-            rooms: "2 rooms",
-            bookingTotal: "5",
-            numberOfUsers: "2",
+          openSnackbar({
+            message: id
+              ? "Coupon successfully updated"
+              : "Coupon successfully created",
+            isError: false,
           })
         );
-        setOpen(false);
       } catch (error) {
       } finally {
         setIsSubmitting(false);
       }
     },
-    [code, type, totalOrPercent, value, customer]
+    [
+      id,
+      code,
+      startDate,
+      endDate,
+      apartmentApplicability,
+      assignedApartments,
+      userApplicability,
+      assignCustomers,
+      type,
+      validity,
+      isReusable,
+      price,
+      percentage,
+      minAmount,
+      maxAmount,
+    ]
   );
 
   return (
-    <form className="w-full flex flex-col gap-5" onSubmit={submitPricing}>
+    <form className="w-full flex flex-col gap-5" onSubmit={createCoupon}>
       <div className=" w-full grid grid-cols-1 gap-5">
         <h5 className=" font-semibold">Add Details</h5>
-        <TextInput
-          inputType="text"
-          isRequired={true}
-          value={code}
-          setValue={setCode}
-          id="coupon-code"
-          placeholder={"Enter Coupon Code"}
-        />
+        {/* coupon code */}
+        <div className=" flex items-center gap-3">
+          <TextInput
+            inputType="text"
+            isRequired={true}
+            value={code}
+            setValue={setCode}
+            id="coupon-code"
+            placeholder={"Enter Coupon Code"}
+          />
+          <div className=" w-fit ">
+            <LoadingButton
+              type="button"
+              isLoading={false}
+              label="Generate Code"
+              clickHandler={generateCouponCode}
+            />
+          </div>
+        </div>
+        {/*  date */}
+        <div className=" grid grid-cols-1  md:grid-cols-2 gap-3">
+          <DateInput
+            inputType="date"
+            isRequired={true}
+            value={startDate}
+            setValue={setStartDate}
+            id="start-date"
+            placeholder="Start date"
+            label="Start date"
+          />
+          <DateInput
+            inputType="date"
+            isRequired={true}
+            value={endDate}
+            setValue={setEndDate}
+            id="end-date"
+            placeholder="End date"
+            label="End date"
+          />
+        </div>
+        {/* apartment select */}
         <Select
           isRequired={true}
-          value={type}
-          setValue={setType}
-          id="coupon-type"
+          value={apartmentApplicability}
+          setValue={setApartmentApplicability}
+          id="apartment-applicability"
         >
           <option value="" disabled>
-            Select Coupon Type
+            Apartment
           </option>
-          <option value="service">Service Coupon</option>
-          <option value="discount">Discount Coupon</option>
+          <option value="all">All</option>
+          <option value="specific">Specific</option>
         </Select>
+        {apartmentApplicability === "specific" && (
+          <Search
+            componentId="apartment"
+            placeholder="Assign to apartments"
+            id="search-apartments"
+            multipleSelect={true}
+          />
+        )}
+        {/* user selection */}
         <Select
           isRequired={true}
-          value={totalOrPercent}
-          setValue={setTotalOrPercent}
-          id="total-or-percent"
+          value={userApplicability}
+          setValue={setUserApplicability}
+          id="user-aplicability"
         >
           <option value="" disabled>
-            Percent or Total
+            User
           </option>
-          <option value="percent">Percent</option>
-          <option value="total">Total</option>
+          <option value="all">All</option>
+          <option value="specific">Specific</option>
         </Select>
-        <TextInput
-          inputType="text"
-          isRequired={true}
-          value={value}
-          setValue={setValue}
-          id="coupon-value"
-          placeholder={"Enter Value of Coupon"}
-        />
-      </div>{" "}
-      <div className=" w-full grid grid-cols-1 gap-5">
-        <h5 className=" font-semibold">Customer</h5>
-        <Search
-          placeholder="Search Customrs to assign to"
-          id="searcg-customer"
-        />
+        {userApplicability === "specific" && (
+          <Search
+            multipleSelect={true}
+            componentId="customer"
+            placeholder="Assign to customer"
+            id="search-customer"
+          />
+        )}
+        <div className=" grid grid-cols-1  md:grid-cols-2 gap-3">
+          {/* type */}
+          <Select
+            isRequired={true}
+            value={type}
+            setValue={setType}
+            id="coupon-type"
+          >
+            <option value="" disabled>
+              Select Coupon Type
+            </option>
+            <option value="price">Price</option>
+            <option value="percentage">Percentage</option>
+          </Select>
+          {type === "price" && (
+            <TextInput
+              inputType="number"
+              isRequired={true}
+              value={price}
+              setValue={setPrice}
+              id="coupon-price"
+              placeholder={"Enter price"}
+            />
+          )}
+          {type === "percentage" && (
+            <TextInput
+              inputType="number"
+              isRequired={true}
+              value={percentage}
+              setValue={setPercentage}
+              id="coupon-percentage"
+              placeholder={"Enter percentage"}
+            />
+          )}
+        </div>
+
+        {/* validity and reusability */}
+        <div className=" grid grid-cols-1  md:grid-cols-2 gap-3">
+          <Select
+            isRequired={true}
+            value={validity}
+            setValue={setValidity}
+            id="validity"
+          >
+            <option value="" disabled>
+              Validity
+            </option>
+            <option value="temporary">Temporary</option>
+            <option value="permanent">Permanent</option>
+          </Select>
+          <Select
+            isRequired={true}
+            value={isReusable}
+            setValue={setIsReusable}
+            id="is-reusable"
+          >
+            <option value="" disabled>
+              Reusable
+            </option>
+            <option value="yes">Yes</option>
+            <option value="no">No</option>
+          </Select>
+        </div>
+
+        {/* max annd minimum amounts */}
+        <div className=" grid grid-cols-1 md:grid-cols-2 gap-5">
+          <TextInput
+            inputType="number"
+            isRequired={false}
+            value={minAmount}
+            setValue={setMinAmount}
+            id="minimum-amount"
+            placeholder={"Minimum amount"}
+          />
+
+          <TextInput
+            inputType="number"
+            isRequired={false}
+            value={maxAmount}
+            setValue={setMaxAmount}
+            id="maximum-amount"
+            placeholder={"Maximum amount"}
+          />
+        </div>
       </div>
       <div className=" flex items-center gap-5 mt-10">
         <LoadingButton