+++ src/components/inputs/search/index.tsx
@@ -1,4 +1,4 @@
-import { ChangeEvent, useCallback, useState } from "react";
+import { ChangeEvent, useCallback, useEffect, useState } from "react";
 import {
   Combobox,
   ComboboxInput,
@@ -6,25 +6,50 @@ import {
   ComboboxOptions,
 } from "@headlessui/react";
 import SearchIcon from "../../../assets/icons/search";
-import customers from "../../../assets/temp-api-mockup-data/customers.json";
 import apartments from "../../../assets/temp-api-mockup-data/apartments.json";
+import CancelIcon from "../../../assets/icons/cancel";
+import useGetAllCustomersLists from "../../../services-hooks/useGetAllCustomersList";
+import LoaderIcon from "../../../assets/icons/loader";
 
 export default function Search({
   id,
   placeholder,
   componentId,
   setValue,
+  multipleSelect,
 }: {
   id: string;
   placeholder: string;
   componentId?: "customer" | "apartment";
   setValue?: Function;
+  multipleSelect?: boolean;
 }) {
-  const [selected, setSelected] = useState(null);
   const [keywords, setKeywords] = useState("");
-  const [filteredResult, setFilteredResult] = useState(
-    componentId === "apartment" ? apartments?.data : customers?.data
-  );
+  const [filteredResult, setFilteredResult] = useState<any>([]);
+  // customers list
+  const {
+    data: customers,
+    isLoading,
+    isFailed,
+    setIsFailed,
+    retryFunction,
+    pagination,
+  } = useGetAllCustomersLists({
+    page: 1,
+    search: componentId === "customer" ? keywords : "",
+  });
+
+  // populate fultered list on data search
+  useEffect(() => {
+    setFilteredResult(
+      componentId === "customer" ? customers : apartments?.data
+    );
+  }, [componentId, customers]);
+
+  const [selectedList, setSelectedList] = useState<
+    { id: string; name: string }[]
+  >([]);
+  const [selected, setSelected] = useState(null);
 
   const handleSearch = useCallback(
     (e: ChangeEvent<HTMLInputElement>) => {
@@ -37,17 +62,16 @@ export default function Search({
             apartment?.location?.toLowerCase()?.includes(value.toLowerCase())
         );
         setFilteredResult(value === "" ? apartments?.data : result);
-      } else {
-        const result = customers.data.filter(
-          (customer) =>
-            customer?.firstname?.toLowerCase()?.includes(value.toLowerCase()) ||
-            customer?.lastname?.toLowerCase()?.includes(value.toLowerCase())
-        );
-        setFilteredResult(value === "" ? customers?.data : result);
       }
     },
     [customers, apartments, componentId]
   );
+
+  const addToList = useCallback((item: { id: string; name: string }) => {
+    if (item?.id) {
+      setSelectedList((prev) => [...prev, item]);
+    }
+  }, []);
 
   const onSelected = useCallback((item: any) => {
     setSelected(item);
@@ -56,9 +80,20 @@ export default function Search({
     }
     if (componentId === "apartment") {
       setKeywords(item ? `${item?.name}` : "");
-    } else {
-      setKeywords(item ? `${item?.firstname} ${item?.lastname}` : "");
+      addToList(item);
+    } else if (componentId === "customer") {
+      setKeywords(item ? `${item?.first_name} ${item?.last_name}` : "");
+      addToList({
+        id: item?.id,
+        name: `${item?.first_name} ${item?.last_name}`,
+      });
     }
+  }, []);
+
+  const removeFromList = useCallback((id: string) => {
+    setSelectedList((prev) => {
+      return prev.filter((item) => item?.id !== id);
+    });
   }, []);
 
   return (
@@ -67,11 +102,30 @@ export default function Search({
       onChange={onSelected}
       // onClose={() => setKeywords("")}
     >
+      {selectedList?.length > 0 && multipleSelect && (
+        <div className=" flex gap-2 flex-wrap">
+          {selectedList.map((item) => (
+            <span
+              key={item?.id}
+              className=" p-2 bg-primary/20 text-primary flex items-center gap-1"
+            >
+              {item?.name}
+              <button onClick={() => removeFromList(item?.id)}>
+                <CancelIcon />
+              </button>
+            </span>
+          ))}
+        </div>
+      )}
       <label
         htmlFor={id}
         className=" p-1 flex items-center gap-2 border rounded-lg text-sm"
       >
-        <SearchIcon />
+        {isLoading ? (
+          <LoaderIcon className=" animate-spin size-6" />
+        ) : (
+          <SearchIcon />
+        )}
         <ComboboxInput
           id={id}
           aria-label="Assignee"
@@ -93,7 +147,7 @@ export default function Search({
           >
             {componentId === "apartment"
               ? item?.name
-              : `${item.firstname} ${item?.lastname}`}{" "}
+              : `${item?.first_name} ${item?.last_name}`}{" "}
           </ComboboxOption>
         ))}
       </ComboboxOptions>