+++ src/services-hooks/useGetCouponLists.tsx
@@ -1,14 +1,25 @@
 import { useCallback, useEffect, useState } from "react";
 import useAxios from "../useHooks/useAxios";
 import { useAppDispatch, useAppSelector } from "../stores/hooks";
-import tempAptData from "../assets/temp-api-mockup-data/coupons.json";
 import {
   updateCouponsList,
   addToPaginationHistory,
 } from "../stores/apiData/coupons-lists";
 
 //axios instace interceptor for access token integration and refresh tokens
-export default function useGetAllCoupons({ page = 1 }: { page?: number }) {
+export default function useGetAllCoupons({
+  page = 1,
+  start_date,
+  end_date,
+  sort = "desc",
+  search = "",
+}: {
+  page?: number;
+  start_date?: string;
+  end_date?: string;
+  sort?: "desc" | "asc" | string;
+  search?: string;
+}) {
   const axios = useAxios();
   const dispatch = useAppDispatch();
   const {
@@ -27,64 +38,69 @@ export default function useGetAllCoupons({ page = 1 }: { page?: number }) {
     from: number;
     to: number;
   }>({} as any);
-
-  const getAllCoupons = useCallback(async () => {
-    setIsLoading(true);
+  const getCouponList = useCallback(async () => {
     try {
+      setIsLoading(true);
       //check store if this requested data has been saved previously and retirve it
       //if not, make a new request and save into store
       const foundPage = store_pagination.find(
         (item) => item?.pagination_data?.current_page === page
       );
-      if (foundPage) {
+      if (
+        foundPage &&
+        !(start_date && end_date) &&
+        !(sort === "asc") &&
+        !search
+      ) {
         setPagination(foundPage?.pagination_data);
         dispatch(updateCouponsList({ data: foundPage?.data }));
       } else {
-        // const response = await axios.post(`/institution-list?page=${page}`);
-        // const responseData = response?.data?.data;
-        // const institutions = response?.data?.institution;
-        dispatch(updateCouponsList({ data: tempAptData.data }));
-
-        // TODO: UPDATE based on backend pagination response
-        //CURRENTLY: Pagination is not being returned for this dataset,
-        //TEMPORARY SOLUTION: Hardcoding pagination
-        // const { data, current_page, last_page, per_page, total, from, to } =
-        //   responseData;
-        // const paginationDataset = {
-        //   current_page: 1,
-        //   last_page: 1,
-        //   per_page: 20,
-        //   total: 4,
-        //   from: 1,
-        //   to: 1,
-        // };
-        // dispatch(update_institution({data}));
+        const response = await axios.get(
+          start_date && end_date
+            ? `/admin/coupon?sort=${sort}&limit=20&search=${
+                search || ""
+              }&page=${page}&start_date=${start_date}&end_date=${end_date}`
+            : `/admin/coupon?sort=${sort}&limit=20&search=${
+                search || ""
+              }&page=${page}`
+        );
+        const { coupon } = response?.data?.data;
+        const { data, current_page, last_page, per_page, total, from, to } =
+          coupon;
+        const paginationDataset = {
+          current_page,
+          last_page,
+          per_page,
+          total,
+          from,
+          to,
+          length: data?.length,
+        };
+        dispatch(updateCouponsList({ data }));
         dispatch(
           addToPaginationHistory({
-            pagination_data: tempAptData.pagination,
-            data: tempAptData.data,
+            pagination_data: paginationDataset,
+            data: data,
           })
         );
-        setPagination(tempAptData.pagination);
+        setPagination(paginationDataset);
       }
       setIsLoading(false);
     } catch (error) {
       setIsFailed(true);
-    } finally {
-      setIsLoading(false);
     }
-  }, [page]);
+  }, [page, start_date, end_date, sort, search]);
 
   useEffect(() => {
-    getAllCoupons();
-  }, [page]);
+    getCouponList();
+  }, [page, start_date, end_date, sort, search]);
 
   return {
     data,
     isLoading,
     isFailed,
     setIsFailed,
-    retryFunction: getAllCoupons,
+    retryFunction: getCouponList,
     pagination,
   };
 }